[Patchset 4] Log delay of task_runner tasks.
Bug: 78765090

+++ include/perfetto/base/task_runner.h
@@ -83,7 +83,7 @@ class TaskRunner {
 
  private:
 #if PERFETTO_DCHECK_IS_ON()
-  std::vector<std::pair<TimeMillis, uint32_t>> delay_histogram_ms_{
+  std::vector<std::pair<TimeMillis, uint64_t>> delay_histogram_ms_{
       {{TimeMillis(5), 0},
        {TimeMillis(10), 0},
        {TimeMillis(50), 0},
